Check for dependencies in stages
Check for all dependencies other than snapraid and snapper
immediately, while continuing to wait until after option parsing
is complete to check for snapraid and snapper in case the user
specifies the --snapper-path and/or --snapraid-path options.

Implemented by making check_dependencies() look for its arguments,
instead of a fixed list, so that it can be called multiple times
with different arguments to check for the appropriate dependencies
at the appropriate time.

Also, instead of ignoring the dependencies only used if -x option
is specified, call check_dependencies() a third time from the
escape_stderr() function, so that those dependencies are checked
iff they are needed.