When reading awk script from heredoc, remove comments/whitespace
Reduce the size of the $awk_script variable by almost 80% by running
a simple awk script on it as it is read from the heredoc, to remove
comments and redundant whitespace.

Since this new awk script that shortens the original awk script only
runs once, while the original awk script is called repeatedly in a
loop, this may be more efficient, and has no real disadvantage. At
worst, it wastes one call to awk and one subshell for the command
substitution, minus a call to read.