Be consistent when redirecting stdout to stderr
Redirect stdout for individual commands rather than entire functions.
Avoids making >&2 a hidden required argument for these functions when
the functions output to stdout, expecting the redirection to occur when
the function is called.

This was already the behavior for all functions except interactive_ask(),
which now behaves consistently with other functions like invalid_argument()