Avoid disabling shellcheck directive SC2064
Use single quotes when defining RETURN trap instead of double quotes.
Makes no difference in behavior because $1 will be the same when the
function returns, but avoids having to disable a shellcheck warning.

Also, add double quotes around $1 inside the single quotes. This
doesn't change behavior since $1 will always be a single word, but
is visually more consistent with the previous line.