new fix -d/--filter-disk option parsing
When doing a fix operation, check to see if the -d option has been
specified. If so, use the live filesystem only for the disk(s)
specified, creating pre/post snapshots, and use the last-synced
snapshots for all other disks, since we know they will not be written
to during the fix operation.