Expand behavior of cleanup command
When the cleanup command is used, delete all snapshots with the
snapraid-btrfs userdata key defined except the one we last completed a
sync against. Previous behavior was to only delete older snapshots, but
there is no benefit to keeping any of the snapshots which aren't
reflected in the parity files.