Expand awk script in parse_snapper_ls() to avoid piping to tr
Accept an optional third argument to parse_snapper_ls() which specifies
the separator to use in between matched snapshots. By default matching
snapshot numbers are separated by newlines as before, if $3 is not
specified. This avoids piping the output of the awk script to tr to
replace the newlines with spaces in snapper_ls().