Revert "Change config file location default behavior to match snapraid"
This reverts commit 053fc37b9df0d1712ec56f002135055cfd1634d2.

The intent of this change was to match the snapraid behavior, but
the snapraid documentation was misleading - /usr/local/etc is only
checked if defined as sysconfdir.

Instead the global readonly variable DEFAULT_CONFIG_FILE can be set,
in lieu of autoconf, or the SNAPRAID_CONFIG_FILE environment variable
can be set.