Work around snapraid.conf files without trailing newline
If there is no newline at the end of the config file, then create a
temporary config file with a newline appended, and use it as a first
stage in generating the config file that will eventually be used.

This avoids bash's read failing due to EOF, and should not affect
the behavior of snapraid's config file parser (see stream.c in the
snapraid source, specifically the sgetline function).