Trap SIGINT in run_snapraid() during a sync operation
If snapraid sync is interrupted with ctrl-C and exits cleanly, it will
return exit status 0, making it appear that a sync has been completed
successfully, causing snapraid-btrfs to update the userdata key to
reflect a successfully completed sync. Therefore, we trap the INT
signal during sync operations, and only update the snapper userdata
key after snapraid sync exits if the trap was not triggered.