Check for invalid input in parse_snapper_ls()
Check to make sure we found columns labelled "#" and "Userdata" when
parsing snapper ls output. This ensures that if the snapper ls output
format changes again in such a way that the method of looking for
columns with these titles no longer works, awk will print an error
message and return failure.