README.md: Remove recommendation to use recent snapper/snapraid version
This recommendation was added when the script was written and was
untested with older versions of snapper/snapraid. However, newer
versions of snapper/snapraid have the possibility to break the script
(as when snapper changed the column order in the output of its ls
command) so it's no longer true they are less likely to break things.