use_snapshot_fix(): Use 0 only as default without --filter-disk
Allow user to specify --use-snapshot in this case, even though without
the --filter-disk option snapraid will try to fix all disks so using a
read-only snapshot doesn't make sense. This makes the behavior more
consistent, and allows snapraid to determine if it's actually being
asked to write to a read-only snapshot and return an error if so.