run_snapraid(): Add error checking
Since this function is called with '&& true' so that it can handle
nonzero snapraid return statuses and pass them to the caller, errexit
and ERR trap will not trigger within the function. So, add || err_trap
after all the commands where we want to abort and exit if they fail.