Fix comment
It's easier to work around lack of a newline at end of file with the
bash read builtin than it is to do this workaround, so the comment as is
doesn't make sense. But it causes problems elsewhere, and with sed there
is no portable workaround (though GNU sed is fine).