Remove 'rollback' command
Snapper has not supported using this command on any config with a
SUBVOLUME other than / since version 0.2.0. While this change was made
to prevent people from breaking their systems by using it on non-root
subvolumes on the same btrfs filesystem as the root subvolume (when
the default subvolume is expected to be mounted as /, the situation the
rollback command is designed for) and wasn't necessary for the case of
non-root subvolumes stored on a different filesystem, as should be the
case with SnapRAID data drives, where it's okay to change the default
subvolume, snapper does not check for this.

The rollback command also isn't really necessary for snapraid-btrfs, and
was mainly created for the sake of completeness. If snapper changes the
behavior of rollback so that it works with subvolumes other than / on
separate filesystems, this commit may be reverted.